2021.09.01 게시판 첨부 파일, 업로드 파일 삭제 기능 구현 완료

+++ src/main/java/com/cloud/porforio/service/BoardServiceImpl.java
@@ -71,7 +71,16 @@ public class BoardServiceImpl implements BoardService{
 
 	@Override
 	public boolean deleteBoard(int bno) {
-		return (mapper.deleteBoardFile(bno) == 1) || (mapper.deleteBoard(bno) == 1);
+		List<BoardFile> fileList = mapper.selectBoardFileList(bno);
+		
+		boolean isRemoveBoardFile = fileUtils.isRemoveBoardFile(fileList);
+		
+		if(isRemoveBoardFile) {
+			return (mapper.deleteBoardFile(bno) == 1) || (mapper.deleteBoard(bno) == 1);
+		}else {
+			return false;
+		}
+	
 	}
 
 	@Override
